+++ lib/aptos/src/provider/models/types.dart
@@ -1,0 +1,2 @@
+export 'fullnode/types.dart';
+export 'graphql/graphql.dart';